feat: move chat support code from app layout to separate component
