+++ frontend/src/container/AppLayout/index.tsx
@@ -4,14 +4,13 @@
 import './AppLayout.styles.scss';
 
 import * as Sentry from '@sentry/react';
-import { Button, Flex, Modal, Typography } from 'antd';
-import updateCreditCardApi from 'api/billing/checkout';
+import { Flex } from 'antd';
 import getLocalStorageKey from 'api/browser/localstorage/get';
 import getUserLatestVersion from 'api/user/getLatestVersion';
 import getUserVersion from 'api/user/getVersion';
 import cx from 'classnames';
+import ChatSupportGateway from 'components/ChatSupportGateway/ChatSupportGateway';
 import OverlayScrollbar from 'components/OverlayScrollbar/OverlayScrollbar';
-import { SOMETHING_WENT_WRONG } from 'constants/api';
 import { IS_SIDEBAR_COLLAPSED } from 'constants/app';
 import { FeatureKeys } from 'constants/features';
 import ROUTES from 'constants/routes';
@@ -22,7 +21,6 @@ import useFeatureFlags from 'hooks/useFeatureFlag';
 import useLicense from 'hooks/useLicense';
 import { useNotifications } from 'hooks/useNotifications';
 import history from 'lib/history';
-import { CreditCard, X } from 'lucide-react';
 import ErrorBoundaryFallback from 'pages/ErrorBoundaryFallback/ErrorBoundaryFallback';
 import {
 	ReactNode,
@@ -35,7 +33,7 @@ import {
 } from 'react';
 import { Helmet } from 'react-helmet-async';
 import { useTranslation } from 'react-i18next';
-import { useMutation, useQueries } from 'react-query';
+import { useQueries } from 'react-query';
 import { useDispatch, useSelector } from 'react-redux';
 import { useLocation } from 'react-router-dom';
 import { Dispatch } from 'redux';
@@ -48,9 +46,6 @@ import {
 	UPDATE_LATEST_VERSION,
 	UPDATE_LATEST_VERSION_ERROR,
 } from 'types/actions/app';
-import { ErrorResponse, SuccessResponse } from 'types/api';
-import { CheckoutSuccessPayloadProps } from 'types/api/billing/checkout';
-import { License } from 'types/api/licenses/def';
 import AppReducer from 'types/reducer/app';
 import { getFormattedDate, getRemainingDays } from 'utils/timeUtils';
 
@@ -72,11 +67,6 @@ function AppLayout(props: AppLayoutProps): JSX.Element {
 	const isDarkMode = useIsDarkMode();
 
 	const { data: licenseData, isFetching } = useLicense();
-	const [activeLicense, setActiveLicense] = useState<License | null>(null);
-
-	const [isAddCreditCardModalOpen, setIsAddCreditCardModalOpen] = useState(
-		false,
-	);
 
 	const isPremiumChatSupportEnabled =
 		useFeatureFlags(FeatureKeys.PREMIUM_SUPPORT)?.active || false;
@@ -84,42 +74,6 @@ function AppLayout(props: AppLayoutProps): JSX.Element {
 	const showAddCreditCardModal =
 		!isPremiumChatSupportEnabled &&
 		!licenseData?.payload?.trialConvertedToSubscription;
-
-	const handleBillingOnSuccess = (
-		data: ErrorResponse | SuccessResponse<CheckoutSuccessPayloadProps, unknown>,
-	): void => {
-		if (data?.payload?.redirectURL) {
-			const newTab = document.createElement('a');
-			newTab.href = data.payload.redirectURL;
-			newTab.target = '_blank';
-			newTab.rel = 'noopener noreferrer';
-			newTab.click();
-		}
-	};
-
-	const handleBillingOnError = (): void => {
-		notifications.error({
-			message: SOMETHING_WENT_WRONG,
-		});
-	};
-
-	const { mutate: updateCreditCard, isLoading: isLoadingBilling } = useMutation(
-		updateCreditCardApi,
-		{
-			onSuccess: (data) => {
-				handleBillingOnSuccess(data);
-			},
-			onError: handleBillingOnError,
-		},
-	);
-
-	const handleAddCreditCard = (): void => {
-		updateCreditCard({
-			licenseKey: activeLicense?.key || '',
-			successURL: window.location.href,
-			cancelURL: window.location.href,
-		});
-	};
 
 	const { pathname } = useLocation();
 	const { t } = useTranslation(['titles']);
@@ -261,13 +215,6 @@ function AppLayout(props: AppLayoutProps): JSX.Element {
 	const [showTrialExpiryBanner, setShowTrialExpiryBanner] = useState(false);
 
 	useEffect(() => {
-		const activeValidLicense =
-			licenseData?.payload?.licenses?.find(
-				(license) => license.isCurrent === true,
-			) || null;
-
-		setActiveLicense(activeValidLicense);
-
 		if (
 			!isFetching &&
 			licenseData?.payload?.onTrial &&
@@ -396,62 +343,7 @@ function AppLayout(props: AppLayoutProps): JSX.Element {
 				</div>
 			</Flex>
 
-			{showAddCreditCardModal && (
-				<>
-					<div className="chat-support-gateway">
-						<Button
-							className="chat-support-gateway-btn"
-							onClick={(): void => setIsAddCreditCardModalOpen(true)}
-						>
-							<svg
-								xmlns="http://www.w3.org/2000/svg"
-								viewBox="0 0 28 32"
-								className="chat-support-gateway-btn-icon"
-							>
-								<path d="M28 32s-4.714-1.855-8.527-3.34H3.437C1.54 28.66 0 27.026 0 25.013V3.644C0 1.633 1.54 0 3.437 0h21.125c1.898 0 3.437 1.632 3.437 3.645v18.404H28V32zm-4.139-11.982a.88.88 0 00-1.292-.105c-.03.026-3.015 2.681-8.57 2.681-5.486 0-8.517-2.636-8.571-2.684a.88.88 0 00-1.29.107 1.01 1.01 0 00-.219.708.992.992 0 00.318.664c.142.128 3.537 3.15 9.762 3.15 6.226 0 9.621-3.022 9.763-3.15a.992.992 0 00.317-.664 1.01 1.01 0 00-.218-.707z" />
-							</svg>
-						</Button>
-					</div>
-
-					{/* Add Credit Card Modal */}
-					<Modal
-						className="add-credit-card-modal"
-						title={<span className="title">Add Credit Card for Chat Support</span>}
-						open={isAddCreditCardModalOpen}
-						closable
-						onCancel={(): void => setIsAddCreditCardModalOpen(false)}
-						destroyOnClose
-						footer={[
-							<Button
-								key="cancel"
-								onClick={(): void => setIsAddCreditCardModalOpen(false)}
-								className="cancel-btn"
-								icon={<X size={16} />}
-							>
-								Cancel
-							</Button>,
-							<Button
-								key="submit"
-								type="primary"
-								icon={<CreditCard size={16} />}
-								size="middle"
-								loading={isLoadingBilling}
-								disabled={isLoadingBilling}
-								onClick={handleAddCreditCard}
-								className="add-credit-card-btn"
-							>
-								Add Credit Card
-							</Button>,
-						]}
-					>
-						<Typography.Text className="add-credit-card-text">
-							You&apos;re currently on{' '}
-							<span className="highlight-text">Trial plan</span>. Add a credit card to
-							access SigNoz chat support to your workspace.
-						</Typography.Text>
-					</Modal>
-				</>
-			)}
+			{showAddCreditCardModal && <ChatSupportGateway />}
 		</Layout>
 	);
 }
